Fix int/size_t type in alpm_list_count() call
alpm_list_count() returns size_t, which we should use to store the
result since it is easy enough to format for printing.

+++ src/pacman/callback.c
@@ -318,9 +318,9 @@ void cb_question(alpm_question_t event, void *data1, void *data2,
 		case ALPM_QUESTION_SELECT_PROVIDER:
 			{
 				alpm_list_t *providers = (alpm_list_t *)data1;
-				int count = alpm_list_count(providers);
+				size_t count = alpm_list_count(providers);
 				char *depstring = alpm_dep_compute_string((alpm_depend_t *)data2);
-				printf(_(":: There are %d providers available for %s:\n"), count,
+				printf(_(":: There are %zd providers available for %s:\n"), count,
 						depstring);
 				free(depstring);
 				select_display(providers);
